Initial check in of Python functionality. Added core/smartbody/python26. Python libs assumed to be in Now using SBxxx classes to extend Character, Joint, Skeleton and provide a cleaner interface. Current python interface can be seen in SbmPython.cpp file. Restore collision geometry to common.sk. Removed reach script from default-init.seq

+++ core/smartbody/smartbody-lib/src/sbm/me_ct_motion_player.h
@@ -51,10 +51,10 @@ public:
 	MeCtMotionPlayer(SbmCharacter* c);
 	~MeCtMotionPlayer();
 
-	void init(SbmPawn* pawn, std::string motionName, int n);
+	void init(SbmPawn* pawn, std::string motionName, double n);
 
-	void setFrameNum(int n);
-	int getFrameNum();
+	void setFrameNum(double n);
+	double getFrameNum();
 
 	void setMotionName(std::string name);
 	std::string getMotionName();
@@ -73,7 +73,7 @@ private:
 	MeController*					controller;		
 	SkChannelArray					channels;
 
-	int								frameNum;
+	double							frameNum;
 	std::string						motionName;
 	bool							isActive;
 };
